pyproject.toml
[tool.poetry]
name = "gql-django"
version = "0.1.0"
description = ""
authors = ["Tra Doan"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
graphene_django = "^3.0.0"
psycopg2 = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
